+++ noodles-bcf/src/async/reader/record.rs
@@ -40,7 +40,7 @@ mod tests {
     use noodles_vcf::{
         header::{format, info},
         record::{
-            genotypes::{genotype::field::Value as GenotypeFieldValue, Keys},
+            genotypes::{values::field::Value as GenotypeFieldValue, Keys},
             info::field::Value as InfoFieldValue,
             Filters as VcfFilters, Genotypes as VcfGenotypes, Ids, Position, QualityScore,
         },